requirements.txt
unstructured>=0.17.2
python-docx>=1.1.2
PyPDF2>=3.0.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
google-generativeai>=0.4.1
requests>=2.31.0
streamlit>=1.25.0
tqdm>=4.67.1
numpy>=1.23.0
nltk>=3.8.1
rank_bm25>=0.2.2
tiktoken>=0.5.1 